+++ .gitignore
@@ -1,2 +1,2 @@
 resourse
-frontend-spa/node_modules
+app/node_modules
