+++ app/README.md
@@ -1,4 +1,4 @@
-# frontend
+# app
 
 > A Vue.js project
