+++ app/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <title>frontend</title>
+    <title>app</title>
   </head>
   <body>
     <div id="app"></div>
