+++ app/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "frontend",
+  "name": "app",
   "version": "1.0.0",
   "description": "A Vue.js project",
   "author": "y_ara",
