+++ server/internal/pkg/config/struct.go
